split rmq in 2 to reflect its role as a driving or driven adapter
