+++ rust/shell/src/rmqpub/job.rs
