+++ rust/shell/src/rmqpub/post.rs
